Use a unified variants cache
