+++ crates/uv-distribution/Cargo.toml
@@ -29,6 +29,7 @@ uv-git = { workspace = true }
 uv-git-types = { workspace = true }
 uv-metadata = { workspace = true }
 uv-normalize = { workspace = true }
+uv-once-map = { workspace = true }
 uv-pep440 = { workspace = true }
 uv-pep508 = { workspace = true }
 uv-platform-tags = { workspace = true }
