+++ crates/uv-distribution/src/distribution_database.rs
@@ -3,13 +3,10 @@ use std::future::Future;
 use std::io;
 use std::path::Path;
 use std::pin::Pin;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 use std::task::{Context, Poll};
 
-use anyhow::anyhow;
-use futures::{FutureExt, StreamExt, TryStreamExt};
-use itertools::Itertools;
-use owo_colors::OwoColorize;
+use futures::{FutureExt, TryStreamExt};
 use rustc_hash::FxHashMap;
 use tempfile::TempDir;
 use tokio::io::{AsyncRead, AsyncSeekExt, ReadBuf};
@@ -26,9 +23,8 @@ use uv_client::{
 use uv_configuration::BuildOutput;
 use uv_distribution_filename::WheelFilename;
 use uv_distribution_types::{
-    BuildableSource, BuiltDist, Dist, DistributionId, HashPolicy, Hashed, Identifier, IndexUrl,
-    InstalledDist, Name, Node, RegistryBuiltDist, RegistryVariantsJson, Resolution, ResolvedDist,
-    ResourceId, SourceDist,
+    BuildableSource, BuiltDist, Dist, HashPolicy, Hashed, IndexUrl, InstalledDist, Name,
+    RegistryVariantsJson, SourceDist,
 };
 use uv_extract::hash::Hasher;
 use uv_fs::write_atomic;
@@ -37,7 +33,6 @@ use uv_pypi_types::{HashDigest, HashDigests};
 use uv_redacted::DisplaySafeUrl;
 use uv_types::{BuildContext, BuildStack, VariantsTrait};
 use uv_variants::resolved_variants::ResolvedVariants;
-use uv_variants::score_variant;
 use uv_variants::variants_json::{VariantPropertyType, VariantsJsonContent};
 
 use crate::archive::Archive;
@@ -590,7 +585,7 @@ impl<'a, Context: BuildContext> DistributionDatabase<'a, Context> {
         let known_properties: Vec<VariantPropertyType> = known_properties.into_iter().collect();
 
         // Compute the set of available variants.
-        // Run all providers.
+        // TODO(charlie): Cache (to disk) by provider.
         let mut resolved_priorities = FxHashMap::default();
         for (name, provider) in &variants_json.providers {
             debug!("Querying provider {name}");
@@ -1289,199 +1284,4 @@ impl LocalArchivePointer {
     pub fn to_cache_info(&self) -> CacheInfo {
         CacheInfo::from_timestamp(self.timestamp)
     }
-}
-
-/// A very simple variants provider cache
-///
-/// TODO(konsti): Cache by provider provider plugin and `requires` compatibility
-/// TODO(konsti): Cache to disk
-#[derive(Debug, Default)]
-pub struct VariantProviderCache {
-    cache: Mutex<FxHashMap<ResourceId, ResolvedVariants>>,
-}
-
-impl VariantProviderCache {
-    pub fn get_resolved_variants(&self, resource_id: &ResourceId) -> Option<ResolvedVariants> {
-        self.cache
-            .lock()
-            .expect("there was a panic in another thread")
-            .get(resource_id)
-            .cloned()
-    }
-
-    pub fn insert_resolved_variants(
-        &self,
-        resource_id: ResourceId,
-        resolved_variants: ResolvedVariants,
-    ) {
-        self.cache
-            .lock()
-            .expect("there was a panic in another thread")
-            .insert(resource_id, resolved_variants);
-    }
-}
-
-/// TODO(konsti): Find a better home for those functions
-pub async fn resolve_variants<Context: BuildContext>(
-    resolution: Resolution,
-    distribution_database: DistributionDatabase<'_, Context>,
-    variants_cache: Arc<VariantProviderCache>,
-) -> anyhow::Result<Resolution> {
-    // Fetch variants.json and then query providers, running in parallel for all dists
-    // TODO(konsti): Reuse in memory index from resolve phase and merge equivalent requests.
-    let dist_resolved_variants: FxHashMap<DistributionId, ResolvedVariants> =
-        futures::stream::iter(
-            resolution
-                .graph()
-                .node_weights()
-                .filter_map(|node| extract_variants(node)),
-        )
-        .map(async |(variants_json, dist)| {
-            let resolved_variants = if let Some(resolved_variants) =
-                variants_cache.get_resolved_variants(&variants_json.resource_id())
-            {
-                resolved_variants.clone()
-            } else {
-                // Fetch variants_json and run providers
-                let resolved_variants = distribution_database
-                    .fetch_and_query_variants(variants_json)
-                    .await?;
-
-                variants_cache.insert_resolved_variants(
-                    variants_json.resource_id(),
-                    resolved_variants.clone(),
-                );
-
-                resolved_variants
-            };
-
-            Ok::<_, anyhow::Error>((dist.distribution_id(), resolved_variants))
-        })
-        // TODO(konsti): Buffer size
-        .buffered(8)
-        .try_collect()
-        .await?;
-
-    // Determine modification to the resolutions to select variant wheels, or error if there
-    // is no matching variant wheel and no matching non-variant wheel.
-    let mut new_best_wheel_index: FxHashMap<DistributionId, usize> = FxHashMap::default();
-    for node in resolution.graph().node_weights() {
-        let Some((_, dist)) = extract_variants(node) else {
-            continue;
-        };
-        let resolved_variants = &dist_resolved_variants[&dist.distribution_id()];
-
-        // Select best wheel
-        let mut highest_priority_variant_wheel: Option<(usize, Vec<usize>)> = None;
-        for (wheel_index, wheel) in dist.wheels.iter().enumerate() {
-            let Some(variant) = wheel.filename.variant() else {
-                // The non-variant wheel is already supported
-                continue;
-            };
-
-            let Some(variants_properties) = resolved_variants.variants_json.variants.get(variant)
-            else {
-                // TODO(konsti): For production, this should be a warning.
-                panic!("Variant {variant} not found in variants.json");
-            };
-
-            let Some(scores) = score_variant(
-                &resolved_variants.variants_json.default_priorities,
-                &resolved_variants.target_variants,
-                variants_properties,
-            ) else {
-                // The wheel is not compatible.
-                continue;
-            };
-
-            if let Some((_, old_scores)) = &highest_priority_variant_wheel {
-                if &scores > old_scores {
-                    highest_priority_variant_wheel = Some((wheel_index, scores));
-                }
-            } else {
-                highest_priority_variant_wheel = Some((wheel_index, scores));
-            }
-        }
-
-        // Determine if we need to modify the resolution
-        if let Some((wheel_index, _scores)) = highest_priority_variant_wheel {
-            debug!(
-                "{} for {}: {}",
-                "Using variant wheel".red(),
-                dist.name(),
-                dist.wheels[wheel_index].filename,
-            );
-            new_best_wheel_index.insert(dist.distribution_id(), wheel_index);
-        } else if dist.best_wheel().filename.variant().is_some() {
-            return Err(anyhow!(
-                "Package {} has no matching wheel for the current platform. Mismatching wheel variants: {}",
-                dist.name(),
-                dist.wheels
-                    .iter()
-                    .filter_map(|wheel| wheel.filename.variant())
-                    .join(", ")
-            ));
-        } else {
-            trace!(
-                "No matching variant wheel, but matching non-variant wheel for {}",
-                dist.name()
-            );
-        }
-    }
-    let resolution = resolution.map(|dist| {
-        let ResolvedDist::Installable {
-            dist,
-            version,
-            variants_json,
-        } = dist
-        else {
-            return None;
-        };
-        let Dist::Built(BuiltDist::Registry(dist)) = &**dist else {
-            return None;
-        };
-        // Check whether there is a matching variant wheel we want to use instead of the default.
-        let best_wheel_index = new_best_wheel_index.get(&dist.distribution_id())?;
-        Some(ResolvedDist::Installable {
-            dist: Arc::new(Dist::Built(BuiltDist::Registry(RegistryBuiltDist {
-                wheels: dist.wheels.clone(),
-                best_wheel_index: *best_wheel_index,
-                sdist: dist.sdist.clone(),
-            }))),
-            variants_json: variants_json.clone(),
-            version: version.clone(),
-        })
-    });
-
-    Ok(resolution)
-}
-
-fn extract_variants(node: &Node) -> Option<(&RegistryVariantsJson, &RegistryBuiltDist)> {
-    let Node::Dist { dist, .. } = node else {
-        // The root node has no variants
-        return None;
-    };
-    let ResolvedDist::Installable {
-        dist,
-        variants_json,
-        ..
-    } = dist
-    else {
-        // TODO(konsti): Installed dists? Or is that not a thing here?
-        return None;
-    };
-    let Some(variants_json) = variants_json else {
-        return None;
-    };
-    let Dist::Built(BuiltDist::Registry(dist)) = &**dist else {
-        return None;
-    };
-    if !dist
-        .wheels
-        .iter()
-        .any(|wheel| wheel.filename.variant().is_some())
-    {
-        return None;
-    }
-    Some((variants_json, dist))
 }
