+++ crates/uv/src/commands/project/sync.rs
@@ -18,7 +18,7 @@ use uv_configuration::{
     Preview, PreviewFeatures, TargetTriple,
 };
 use uv_dispatch::BuildDispatch;
-use uv_distribution::{DistributionDatabase, VariantProviderCache, resolve_variants};
+use uv_distribution::{DistributionDatabase, resolve_variants};
 use uv_distribution_types::{
     DirectorySourceDist, Dist, Index, Requirement, Resolution, ResolvedDist, SourceDist,
 };
@@ -712,7 +712,6 @@ pub(super) async fn do_sync(
         DistributionDatabase::new(&client, &build_dispatch, concurrency.downloads);
 
     // Read the lockfile.
-    let variants_cache = Arc::new(VariantProviderCache::default());
     let resolution = target
         .to_resolution(
             &marker_env,
@@ -722,7 +721,7 @@ pub(super) async fn do_sync(
             build_options,
             &install_options,
             distribution_database,
-            variants_cache.clone(),
+            state.index().variant_priorities(),
         )
         .await?;
 
@@ -775,7 +774,12 @@ pub(super) async fn do_sync(
     // TODO(konsti): Pass this into operations::install
     let distribution_database =
         DistributionDatabase::new(&client, &build_dispatch, concurrency.downloads);
-    let resolution = resolve_variants(resolution, distribution_database, variants_cache).await?;
+    let resolution = resolve_variants(
+        resolution,
+        distribution_database,
+        state.index().variant_priorities(),
+    )
+    .await?;
 
     let site_packages = SitePackages::from_environment(venv)?;
